fix: getColTitle util
